Add workflow for uploading lab archives
On every push on the main branch, check if any files related to a lab
have changed. If so, create an archive with all files related to that
lab and push it to the `lab-archives` branch.

The `gen-zip.py` script will parse the `config.yaml` file, check the
last commit, and if the commit changes any file related to one of the
labs, extract all files related to that lab and create a zip archive.

Signed-off-by: Stefan Jumarea <stefanjumarea02@gmail.com>