Update lab archive generator workflow
Remove `.zip` from gitignore, this is needed because the workflow will
overwrite the changes to the archive if the file is ignored by git.
Create a commit for removing the outdated archives, since `git
stash pop` will generate a merge conflict otherwise.
Check if the name of the task contains `drills`. If it does not, do not
add it to the archive.

Signed-off-by: Stefan Jumarea <stefanjumarea02@gmail.com>